pyproject.toml
[project]
name = "nhp_ats_tui"
version = "0.1.0"
description = "TUI to edit NHP model-run entities in Azure Table Storage"
authors = [
    { name = "Matt Dray", email = "matt.dray@nhs.net" },
    { name = "YiWen Hon", email = "yiwen.hon1@nhs.net" },
]
requires-python = ">=3.10"

dependencies = [
  "azure-data-tables>=12.4",
  "azure-identity>=1.15",
  "inquirerpy>=0.3.4",
]

[project.scripts]
edit-runs = "nhp_ats_tui.tui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
